version:0.0.2, complete some todos

+++ main.go
@@ -17,21 +17,25 @@ import (
 	"github.com/fatih/color"
 )
 
-var version = "v0.0.1"
+var version = "v0.0.2"
 var dialTimeout = 5 * time.Second
-var defaultIdleTimeout = 1 * time.Minute
+var defaultIdleTimeout = 876000 * time.Hour // 100 years
 
 var help = flag.Bool("h", false, "Show this help doc")
 var ver = flag.Bool("version", false, "Show current version of ondict")
 var word = flag.String("q", "", "Specify the word that you want to query")
-var easyMode = flag.Bool("e", false, "True to show only 'frequent' meaning")
+
+// var easyMode = flag.Bool("e", false, "True to show only 'frequent' meaning")
 var dev = flag.Bool("d", false, "If specified, a static html file will be parsed, instead of an online query, just for dev debugging")
 var verbose = flag.Bool("v", false, "Show debug logs")
 var interactive = flag.Bool("i", false, "Launch an interactive CLI app")
 var server = flag.Bool("serve", false, "Serve as a HTTP server, default on UDS, for cache stuff, make it quicker!")
 var idleTimeout = flag.Duration("listen.timeout", defaultIdleTimeout, "Used with '-serve', the server will automatically shut down after this duration if no new requests come in")
+var listenAddr = flag.String("listen", "", "Used with '-serve', address on which to listen for remote connections. If prefixed by 'unix;', the subsequent address is assumed to be a unix domain socket. Otherwise, TCP is used.")
 var remote = flag.String("remote", "", "Connect to a remote address to get information, 'auto' means it will try to launch a request by UDS. If no local server is working, a new server will be created, with -listen.timeout 1 min.")
 var colour = flag.Bool("color", false, "This flags controls whether to use colors.")
+var render = flag.String("f", "", "render format, 'md' (for markdown, only for mdx engine now), or 'html'")
+var engine = flag.String("e", "", "query engine, 'mdx' or others(online query)")
 
 var mu sync.Mutex // owns history
 var history map[string]string = make(map[string]string)
@@ -43,8 +47,8 @@ func init() {
 	if err != nil {
 		log.Fatal(err)
 	}
-	config := filepath.Join(home, ".config")
-	dataPath = filepath.Join(config, "ondict")
+	configPath := filepath.Join(home, ".config")
+	dataPath = filepath.Join(configPath, "ondict")
 	historyFile = filepath.Join(dataPath, "history.json")
 	if dataPath == "" || historyFile == "" {
 		log.Fatalf("empty datapath/historyfile: %v||%v", dataPath, historyFile)
@@ -59,7 +63,14 @@ func main() {
 	}
 	if !*verbose {
 		log.SetOutput(io.Discard)
+		separatorOpen, separatorClose = "", ""
 	}
+	loadConfig()
+
+	if *render != "md" {
+		gbold, gitalic = "", ""
+	}
+
 	if *ver {
 		fmt.Printf("ondict %s %s %s with %s\n", version, runtime.GOOS, runtime.GOARCH, runtime.Version())
 		return
@@ -70,6 +81,7 @@ func main() {
 	}
 
 	if *interactive {
+		ldoceDict = loadDecodedMdx(filepath.Join(dataPath, "dicts", ldoceMdx)) // TODO(ch): lazy loading for performance?
 		startLoop()
 		return
 	}
@@ -77,18 +89,24 @@ func main() {
 	if *server {
 		stop := make(chan error)
 		p := new(proxy)
-		p.timeout = time.NewTimer(*idleTimeout)
-		dp, err := os.Executable()
-		if err != nil {
-			log.Fatalf("getting ondict path error: %v", err)
+		if *idleTimeout > 0 {
+			p.timeout = time.NewTimer(*idleTimeout)
 		}
-		network, addr := autoNetworkAddressPosix(dp, "")
-		if _, err := os.Stat(addr); err == nil {
-			if err := os.Remove(addr); err != nil {
-				log.Fatalf("removing remote socket file: %v", err)
+		network, addr := ParseAddr(*listenAddr)
+		if network == "auto" || addr == "" {
+			dp, err := os.Executable()
+			if err != nil {
+				log.Fatalf("getting ondict path error: %v", err)
+			}
+			network, addr = autoNetworkAddressPosix(dp, "")
+			if _, err := os.Stat(addr); err == nil {
+				if err := os.Remove(addr); err != nil {
+					log.Fatalf("removing remote socket file: %v", err)
+				}
 			}
 		}
-		log.Printf("%s, start a new server: %s", dp, addr)
+		log.Printf("start a new server: %s/%s/%s/%s", network, addr, *render, *engine)
+		ldoceDict = loadDecodedMdx(filepath.Join(dataPath, "dicts", ldoceMdx)) // TODO(ch): lazy loading for performance?
 		l, err := net.Listen(network, addr)
 		if err != nil {
 			log.Fatal("bad Listen: ", err)
@@ -122,7 +140,7 @@ func main() {
 		netConn, err := net.DialTimeout(network, address, dialTimeout)
 
 		if err == nil { // detect an exsitng server, just forward a request
-			if err := request(netConn); err != nil {
+			if err := request(netConn, *engine, *render); err != nil {
 				log.Fatal(err)
 			}
 			return
@@ -142,6 +160,9 @@ func main() {
 		}
 		args := []string{
 			"-serve=true",
+			"-listen.timeout=2m",
+			"-e=" + *engine,
+			"-f=" + *render,
 		}
 		log.Printf("starting remote: %v", args)
 		if err := startRemote(dp, args...); err != nil {
@@ -154,7 +175,7 @@ func main() {
 			startDial := time.Now()
 			netConn, err = net.DialTimeout(network, address, dialTimeout)
 			if err == nil {
-				if err := request(netConn); err != nil {
+				if err := request(netConn, *engine, *render); err != nil {
 					log.Fatal(err)
 				}
 				return
@@ -168,7 +189,7 @@ func main() {
 		}
 		os.Exit(3)
 	} else if *remote != "" {
-		log.Fatal("TODO: specify a remote address not supported yet")
+		log.Fatal("TODO(ch): specify a remote address not supported yet")
 	}
 
 	// just for offline test.
@@ -181,10 +202,26 @@ func main() {
 		fmt.Println(parseHTML(fd))
 		return
 	}
+
+	if *engine == "mdx" {
+		// io.Copy(os.Stdout, fd)
+		ldoceDict = loadDecodedMdx(filepath.Join(dataPath, "dicts", ldoceMdx)) // TODO(ch): lazy loading for performance?
+		fmt.Println(queryMDX(*word, *render))
+		return
+	}
 	fmt.Println(queryByURL(*word))
 }
 
-func query(word string) string {
+func query(word string, e string, f string) string {
+	if e == "" {
+		e = *engine
+	}
+	if f == "" {
+		f = *render
+	}
+	if e == "mdx" {
+		return queryMDX(word, f)
+	}
 	var res string
 	mu.Lock()
 	if ex, ok := history[word]; ok {
